# docs/api-reference/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/{network}/fee-bump/sponsor": {
      "post": {
        "operationId": "FeeBumpController_sponsor",
        "parameters": [
          {
            "name": "network",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "testnet",
                "mainnet"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeeBumpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The transaction wrapped in a fee-bump envelope and signed (base64 XDR).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeBumpSponsorResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid transaction or disallowed structure.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "invalid_network",
                                "missing_transaction",
                                "invalid_transaction_xdr",
                                "inner_fee_not_zero",
                                "operation_not_sponsorable",
                                "fee_bump_exceeds_cap",
                                "invalid_body"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "unauthorized"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "413": {
            "description": "The request body is larger than 100 KB.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "payload_too_large"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "415": {
            "description": "The request body is not UTF-8 JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "unsupported_media_type"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded for this key.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "rate_limited"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Sponsored-fee flow not configured on this server.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "fee_bump_not_configured"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Sponsor a wind-down transaction's fee for a reserve-locked account.",
        "tags": [
          "fee-bump"
        ]
      }
    },
...
  },
  "info": {
    "title": "LumenWipe API",
    "description": "Programmatic close-out of Stellar accounts. Request bodies must be JSON objects of at most 100 KB: a larger body is rejected with 413 payload_too_large, a non-JSON content type with 415 unsupported_media_type, and a body that is not a JSON object with 400 invalid_body.",
    "version": "0.1.0",
    "contact": {
      "name": "LumenWipe",
      "url": "https://lumenwipe.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "servers": [
    {
      "url": "https://api.lumenwipe.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Integrator API key"
      },
      "integrator-session": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "Short-lived wallet session token from POST /integrator/auth/session"
      }
    },
    "schemas": {
...
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "account_not_found",
              "account_read_failed",
              "account_too_large",
              "admin_api_not_configured",
              "allowances_read_failed",
              "api_key_not_found",
              "aquarius_trustline_missing",
              "backstop_emissions_unclaimed",
              "backstop_token_unconvertible",
              "backstop_withdrawal_cooling_down",
              "backstop_withdrawal_not_queued",
              "bad_request",
              "blend_emissions_trustline_missing",
              "blend_repay_asset_balance_unknown",
              "blend_repay_asset_missing",
              "confirmation_timeout",
              "conversion_floor_missing",
              "conversion_provider_unrecognized",
              "defi_exit_blocked",
              "defi_exit_unsupported",
              "defi_position_unrecognized",
              "defi_positions_blocked",
              "defi_positions_stale",
              "defi_positions_unavailable",
              "destination_not_acknowledged",
              "fee_bump_exceeds_cap",
              "fee_bump_not_configured",
              "forbidden",
              "forward_amount_exceeds_balance",
              "inner_fee_not_zero",
              "integrator_api_not_configured",
              "internal_error",
              "invalid_address",
              "invalid_addresses",
              "invalid_body",
              "invalid_challenge",
              "invalid_decisions",
              "invalid_destination",
              "invalid_memo",
              "invalid_network",
              "invalid_owner",
              "invalid_rate_limit",
              "invalid_request",
              "invalid_signature",
              "invalid_signed_xdr",
              "invalid_source",
              "invalid_spender",
              "invalid_token",
              "invalid_tokens",
              "invalid_transaction_xdr",
              "invalid_tx_hash",
              "key_limit_reached",
              "mediator_key_misconfiguration",
              "mediator_not_configured",
              "memo_required",
              "merged_account_not_found",
              "missing_parameters",
              "missing_transaction",
              "needs_decisions",
              "not_found",
              "operation_not_sponsorable",
              "owner_not_found",
              "path_lookup_failed",
              "payload_too_large",
              "phoenix_trustline_missing",
              "plan_failed",
              "provider_response_unusable",
              "quote_drifted",
              "rate_limited",
              "registry_expired",
              "request_failed",
              "revoke_build_failed",
              "revoke_needs_restore",
              "revoke_simulation_failed",
              "revoke_unsafe",
              "service_unavailable",
              "signer_normalization_unsafe",
              "soroban_token_conversion_failed",
              "soroban_token_conversion_unavailable",
              "soroban_token_conversion_unsafe",
              "soroban_token_needs_restore",
              "soroban_token_route_lost",
              "soroban_token_transfer_failed",
              "soroban_token_transfer_unsafe",
              "soroban_token_unreadable",
              "soroswap_trustline_missing",
              "stats_unavailable",
              "submit_failed",
              "submit_rejected",
              "too_many_addresses",
              "transaction_structure_not_allowed",
              "transactions_failed",
              "transfer_destination_missing",
              "transfer_destination_unusable",
              "trustline_cannot_be_left",
              "trustline_deauthorized_with_balance",
              "tx_not_verified",
              "unauthorized",
              "unprocessable_entity",
              "unsupported_media_type",
              "unsupported_memo_type",
              "vault_undercollateralized",
              "withdraw_before_repay"
            ],
            "description": "Stable machine-readable code. Branch on this, never on the message."
          },
          "message": {
            "type": "string",
            "description": "Plain-language explanation. May be reworded."
          },
          "details": {
            "type": "object",
            "description": "Extra structured context, present only when there is more to say."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "FeeBumpRequestDto": {
        "type": "object",
        "properties": {
          "transaction": {
            "type": "string",
            "description": "The unsigned or user-signed wind-down transaction to sponsor (base64 XDR), inner fee set to zero.",
            "example": "AAAAAgAAAAB..."
          }
        },
        "required": [
          "transaction"
        ]
      },
      "FeeBumpSponsorResponseDto": {
        "type": "object",
        "properties": {
          "transaction": {
            "type": "string",
            "description": "The fee-bumped transaction envelope (base64 XDR), signed by the sponsor.",
            "example": "AAAABQAAAAB..."
          }
        },
        "required": [
          "transaction"
        ]
      },
...
    }
  }
}